provide ios device and simulator stand-alone versions for tvOS framework. Improved the OneAudioPlayer component.
