+++ .idea/xcode.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="XcodeMetaData" PROJECT_FILE="$PROJECT_DIR$/AVPlayerDemo.xcodeproj" />
+</project>
