+++ tvOS/appletvos/tvOSAVPlayerTouch.framework/Headers/tvOSAVPlayerTouch.h
@@ -1,0 +1,23 @@
+//
+//  tvOSAVPlayerTouch.h
+//  tvOSAVPlayerTouch
+//
+//  Created by apple on 15/12/6.
+//  Copyright © 2015年 apple. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+//! Project version number for tvOSAVPlayerTouch.
+FOUNDATION_EXPORT double AVPlayerTouchVersionNumber;
+
+//! Project version string for tvOSAVPlayerTouch.
+FOUNDATION_EXPORT const unsigned char AVPlayerTouchVersionString[];
+
+#import <tvOSAVPlayerTouch/AVPErrors.h>
+#import <tvOSAVPlayerTouch/AVPLicense.h>
+#import <tvOSAVPlayerTouch/OneAudioPlayer.h>
+#import <tvOSAVPlayerTouch/NativeAVPlayerController.h>
+#import <tvOSAVPlayerTouch/FFAVPlayerController.h>
+#import <tvOSAVPlayerTouch/FFAVParser.h>
+#import <tvOSAVPlayerTouch/CustomizedAVSource.h>
