--- tvOSAVPlayerTouch.framework/Modules/module.modulemap
@@ -1,6 +1,0 @@
-framework module tvOSAVPlayerTouch {
-  umbrella header "tvOSAVPlayerTouch.h"
-
-  export *
-  module * { export * }
-}
