--- tvOSAVPlayerTouch.framework/_CodeSignature/CodeResources
@@ -1,183 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>files</key>
-	<dict>
-		<key>Headers/AVPErrors.h</key>
-		<data>
-		QwUKWirZf5+HsE1PLKUYo5+kAZA=
-		</data>
-		<key>Headers/AVPLicense.h</key>
-		<data>
-		8DKZUk4xTAmaeoUh39FOcDgPHeo=
-		</data>
-		<key>Headers/CustomizedAVSource.h</key>
-		<data>
-		eaUcmXKYd0Z+bBLPyUUJFZpax58=
-		</data>
-		<key>Headers/FFAVParser.h</key>
-		<data>
-		qH4it2Rj3ui+LdW0mU50QJDM2hw=
-		</data>
-		<key>Headers/FFAVPlayerController.h</key>
-		<data>
-		Uji+183hGgbM2sZzFvYLdjmK5yQ=
-		</data>
-		<key>Headers/NativeAVPlayerController.h</key>
-		<data>
-		PVxpxRCkq0ImYiYHdY4GapQ1xDE=
-		</data>
-		<key>Headers/OneAudioPlayer.h</key>
-		<data>
-		kABoUFqUj/ii7w7jtOSp598ZVy4=
-		</data>
-		<key>Headers/tvOSAVPlayerTouch.h</key>
-		<data>
-		KJotrGfKObPaGRomUucpinng9EA=
-		</data>
-		<key>Info.plist</key>
-		<data>
-		iJFnMo77DijhQm5LnM34V0gFnbQ=
-		</data>
-		<key>Modules/module.modulemap</key>
-		<data>
-		4bPsfC13rp78t7XLGXpZ5w0rkKk=
-		</data>
-	</dict>
-	<key>files2</key>
-	<dict>
-		<key>Headers/AVPErrors.h</key>
-		<data>
-		QwUKWirZf5+HsE1PLKUYo5+kAZA=
-		</data>
-		<key>Headers/AVPLicense.h</key>
-		<data>
-		8DKZUk4xTAmaeoUh39FOcDgPHeo=
-		</data>
-		<key>Headers/CustomizedAVSource.h</key>
-		<data>
-		eaUcmXKYd0Z+bBLPyUUJFZpax58=
-		</data>
-		<key>Headers/FFAVParser.h</key>
-		<data>
-		qH4it2Rj3ui+LdW0mU50QJDM2hw=
-		</data>
-		<key>Headers/FFAVPlayerController.h</key>
-		<data>
-		Uji+183hGgbM2sZzFvYLdjmK5yQ=
-		</data>
-		<key>Headers/NativeAVPlayerController.h</key>
-		<data>
-		PVxpxRCkq0ImYiYHdY4GapQ1xDE=
-		</data>
-		<key>Headers/OneAudioPlayer.h</key>
-		<data>
-		kABoUFqUj/ii7w7jtOSp598ZVy4=
-		</data>
-		<key>Headers/tvOSAVPlayerTouch.h</key>
-		<data>
-		KJotrGfKObPaGRomUucpinng9EA=
-		</data>
-		<key>Modules/module.modulemap</key>
-		<data>
-		4bPsfC13rp78t7XLGXpZ5w0rkKk=
-		</data>
-	</dict>
-	<key>rules</key>
-	<dict>
-		<key>^</key>
-		<true/>
-		<key>^.*\.lproj/</key>
-		<dict>
-			<key>optional</key>
-			<true/>
-			<key>weight</key>
-			<real>1000</real>
-		</dict>
-		<key>^.*\.lproj/locversion.plist$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>1100</real>
-		</dict>
-		<key>^version.plist$</key>
-		<true/>
-	</dict>
-	<key>rules2</key>
-	<dict>
-		<key>.*\.dSYM($|/)</key>
-		<dict>
-			<key>weight</key>
-			<real>11</real>
-		</dict>
-		<key>^</key>
-		<dict>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-		<key>^(.*/)?\.DS_Store$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>2000</real>
-		</dict>
-		<key>^(Frameworks|SharedFrameworks|PlugIns|Plug-ins|XPCServices|Helpers|MacOS|Library/(Automator|Spotlight|LoginItems))/</key>
-		<dict>
-			<key>nested</key>
-			<true/>
-			<key>weight</key>
-			<real>10</real>
-		</dict>
-		<key>^.*</key>
-		<true/>
-		<key>^.*\.lproj/</key>
-		<dict>
-			<key>optional</key>
-			<true/>
-			<key>weight</key>
-			<real>1000</real>
-		</dict>
-		<key>^.*\.lproj/locversion.plist$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>1100</real>
-		</dict>
-		<key>^Info\.plist$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-		<key>^PkgInfo$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-		<key>^[^/]+$</key>
-		<dict>
-			<key>nested</key>
-			<true/>
-			<key>weight</key>
-			<real>10</real>
-		</dict>
-		<key>^embedded\.provisionprofile$</key>
-		<dict>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-		<key>^version\.plist$</key>
-		<dict>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-	</dict>
-</dict>
-</plist>
